Handle Vault.retrieve_inventory return type
retrieve_inventory was documented to return a Job object but in fact
returns a job ID string. I see that the documentation has been fixed in
master.

Call vault.get_job to translate the returned string instead. This will
work for versions of boto from before boto was fixed, too.

This fixes a crash with "glacier vault sync --wait".